python-app/requirements.txt
numpy==1.24.3
pandas==2.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
tqdm==4.66.1
pyarrow==14.0.0
matplotlib==3.10.7
seaborn==0.13.2
scikit-learn==1.3.2